data-science-project/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
jupyter
langgraph
langchain
langchain-openai
langchain-community
openai
python-dotenv
requests
networkx
ipython